Modified table view of search with code refactor
